flowsint-face-recognition/requirements.txt
# Core dependencies
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1

# AI/ML
insightface==0.7.3
onnxruntime==1.20.1
opencv-python==4.10.0.84
numpy==1.26.4
scikit-learn==1.6.0

# LLM
google-generativeai==0.8.3

# OSINT (will integrate Social Mapper separately)
requests==2.32.3
beautifulsoup4==4.12.3
Pillow==11.0.0

# Geolocation
pyexiftool==0.5.6
geopy==2.4.1

# Database
redis==5.2.0
psycopg2-binary==2.9.10
sqlalchemy==2.0.36

# Utilities
python-multipart==0.0.20
aiofiles==24.1.0
python-dotenv==1.0.1
